Option `--url` is not mandatory
In a recent patch we replaced `os.Exit(...)` with the Cobra error
handling mechanism. A side effect of that change is that the `--url`
option was marked as mandatory. This makes the application fail if it
isn't explicity used even if it has a default value. This patch changes
the application so that it isn't mandatory. Instead we just check that
it isn't empty.

Related: https://github.com/openshift-online/uhc-cli/issues/61
Signed-off-by: Juan Hernandez <jhernand@redhat.com>