Update to SDK 0.1.36
The more relevant change in the new version of the SDK is that the
`cluster.craetor` field has been removed. This means that we need to
update the `cluster describe` command to get the creator following the
link to the subscription and then to the account.

Related: https://jira.coreos.com/browse/SDA-1395
Signed-off-by: Juan Hernandez <jhernand@redhat.com>