Rename `build_deploy.sh` to `build_tag.sh`
The `build_deploy.sh` script was initially created to automate releases,
but at the time there was no way in the `app-interface` project to
trigger a job when a tag is pushed. That has been now added and the
default script name is `build_tag.sh`. This patch renames the script
accordingly.

Signed-off-by: Juan Hernandez <jhernand@redhat.com>