Retrieve all enabled versions
Currently the `cluster create` command retrieves only the first page of
enabled versions. This works fine in environments with less than 100
versions, but in the integration environment there are already 116
versions. This results in rejecting the command when the user tries to
use any of the newer versions that aren't in the first page. To avoid
that this patch changes the command so that it will always retrieve all
the versions, even if there are more than 100.

Signed-off-by: Juan Hernandez <jhernand@redhat.com>