Don't require `HOME` environment variable
This patch changes the tool so that it doesn't require the `HOME`
environment variable to find the home directory of the user, as that
doesn't work in Windows. Instead of that the tool uses now the
`github.com/mitchellh/go-homedir` which works in all the supported
operating systems.

Related: https://jira.coreos.com/browse/SDA-1597
Signed-off-by: Juan Hernandez <jhernand@redhat.com>