Fix empty edit default machine pool
This bug was caused because we use the same func to create and update the nodes builder for a cluster.
It would only update if computes nodes != 0 (which means it was set by the user).
However, the default for edit additional machine pool is -1, because 0 is a valid value.
Since we now use the same `edit machinepool` command for default and additional,
I changed the check to `if computeNodes > 0` which captures both cases.
Related: https://issues.redhat.com/browse/SDA-3709

This also includes a documentation fix for --replicas flag.
Related: https://issues.redhat.com/browse/SDA-3710