Use pager command for `list clusters`
This patch changes the `list clusters` command so that it uses the
a pager command (if available) to display the results page by page.

This will not be enabled by default. To enable it set the `pager` config
variable to the command that you wish to use. For example, to use the
`less -F` command:

```
$ ocm config set pager 'less -F'
```

Related: https://issues.redhat.com/browse/SDA-4280
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>