Add table printer
This patch adds a new `output.Table` type that simplifies writing
tabular ouptut. It is created from the `output.Printer` type introduced
in a previous patch:

```go
// Create the output table:
table, err := printer.NewTable().
        Name("clusters").
        Columns(args.columns).
        Build(ctx)
if err != nil {
        return err
}
```

When a table like this is created it reads the descriptions of the
columns from the corresponding file in the `pkg/output/tables`
directory. In this case it will read the `clusters.yaml` file. That file
should contain the descriptions of the columns of that table, in
particular it should contain the preferred width of the columns. For
example:

```yaml
columns:
- name: id
  header: ID
  width: 32
- name: name
  header: NAME
  width: 28
  ...
- name: external_id
  header: EXTERNAL ID
  width: 36
```

In this example the `external_id` column is defined with a preferred
width of 36 charaters. The table object will use that information to
format the output accordingly. For example, a command like this:

```
$ ocm list clusters --columns id,external_id,name --managed
```

Will result in the following output:

```
ID                                EXTERNAL_ID                   NAME
1i6anlpc6278j4ok5nbp8ik69dl6h53j  9f2c1195-1f1b-4780-ae09-00f6  osde2e-hwlrn
1ku25di05s69u46bbnc0abihl9ig2e9a  155beb89-feec-4246-aab9-1592  osde2e-hxhfh
1lfklioadcqr5oeitfokqkavq4gir25l  e2eb5d92-f0d7-4077-b7b3-a2d8  osde2e-kp4sk
1lfklji54fhu3b64us3a62d9t9bju2r8  NONE                          osde2e-wfw97
1lfkljjopet5vkpnks33j65kb7pilao3  3e567572-f4fb-49cf-b25e-3196  osde2e-o9z8x
...
```

Note that this patch only modifies the `list clusters` command. Other
commands that write tables are still using the previous infrastructure.
That will be addressed in separate patches.

Related: https://issues.redhat.com/browse/SDA-4246
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>