Add column width learning
This patch changes the `output.Table` object so that it accumulates the
first rows of the output (one hundred by default) in order to learn how
to display the values.

Currently that learning is very basic: it adjusts the width of the
columns to not waste space, and only does it for columsn that don't have
a width explicitly defined in the corresponding YAML file.

This works well usually, but may product wrong results if the data in
the first one hundred rows is very different from the data in the
following rows.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>