User friendly message when offline token is no longer valid
This patch changes the CLI so that it will write the following message
when an operation fails because the offline access token is no longer
valid:

```
Offline access token is no longer valid. Go to
https://console.redhat.com/openshift/token to get a new
one and then use the 'ocm login --token=...' command to
log in with that new token.
```

Related: https://issues.redhat.com/browse/SDA-4686
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>