Fix publish release action
The action that publishes releases was uploading artifacts using the
`--data` option of the `curl` tool. That results in processing them as
ASCII and limits the size to approximately 4 MiB. The end result is that
the uploaded binaries don't work. This patch fixes that using the
`--data-binary` option instead.

Related: https://github.com/openshift-online/ocm-cli/issues/319
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>