Fix binary names inside `.sha256` files
Since the introduction of the `publish-release` action the `.sha256`
files contain incorrect file names. For example, the name of the actual
released binary for Linux and x86_64 is `ocm-linux-amd64` but the file
name inside the `.sha256` file is just `ocm`. This patch fixes that
issue.

Related: https://github.com/openshift-online/ocm-cli/issues/322
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>