OCM-1888: Add docs for ocm delete account subcommands and arguments
Prevously when using the help in OCM-CLI it did not list all available sub commands.
This was due to how we currently alias urls in the CLI. Now added documentation about using the
urls(path) or provide a resource_id for the resource to be deleted. Another effort has been started
to understand and standardize the help format, which in turn will help in the task OCM-5179.
The template format has been used from https://pkg.go.dev/text/template.

Signed-off-by: Chetan Giradkar <cgiradka@redhat.com>