Adds tech preview warning to override url
Allows override URL but when running with the debug flag this will
inform users that this functionality is tech preview and may have issues
while we build and test this.

This should allow SRE to be able to use this feature immediately
following the next release while setting the expectation with customers
who may find this that it may still have some unexpected quirks, and to
fall back to supported methodology for using this tool.